move aria-at route to app.js
